--- src/Pages/Education.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-class Education extends React.Component {
-  render() {
-    return <h1>학술행사/교육 페이지</h1>;
-  }
-}
-
-export default Education;
