--- src/Pages/Home.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function Main(props) {
-  return (
-    <>
-      <h3> 안녕하세요. 메인페이지 입니다.</h3>
-    </>
-  );
-}
-
-export default Main;
