--- src/Pages/Information.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-class Info extends React.Component {
-  render() {
-    return <h1>정보광장 페이지</h1>;
-  }
-}
-
-export default Info;
