--- src/Pages/Journal.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-class Journal extends React.Component {
-  render() {
-    return <h1>학회지 페이지</h1>;
-  }
-}
-
-export default Journal;
